File picker undo should work
It turns out that updates made to the the input value programatically are not added to the undo buffer. Any time the user makes a change to the input box value, we need to make sure they don't have to delete the programatically added auto complete value because otherwise the undo buffer will contain their delete but not the programatic add. This results in unusable undo. Fixes #69536